Crate calendar_mirror

A simple Rust project to show the events of a user's google calendar on a PaPiRus ePaper display on a Raspberry Pi.

Macros

cloneable
Newtype implementation for Clone types.

copyable
Newtype implementation for Copy types.

err
Macro that creates an enum wrapper with a variant corresponding to each of a series of either encapsulated or new errors.

non_copyable
Newtype implementation for non-Copy types.

reffable
Newtype implementation for non-Clone types.

stm
Create a state machine

Structs

PathError
Error returned if path to systemd service file isn't valid UTF-8

Enums

Error
Overarching error type that encapsulates all errors in the application