Macro calendar_mirror::err

macro_rules! err {
    ($(#[$outer:meta])* $enum_name:ident { $( $var:ident($embedded:path) ),+ } ) => { ... };
}

Macro that creates an enum wrapper with a variant corresponding to each of a series of either encapsulated or new errors.

The macro also implements the From trait for the enum allowing the user to convert one of the encapsulated errors (or indeed any arbitrary payload where the user is creating a new error) into one of the enum variants.

Examples

use std::fs::File;
use io::Error;

#[derive(Clone, Debug)]
pub struct Pin(pub usize);

err!(Error {
    #[doc="Errors can that can be generated when reading GPIOs"]
    File(io::Error),
    InvalidPin(Pin)
});

type GPIOResult<T>=Result<T, Error>

Instantiate a Pin error as follow:

 let pin_error=Error::InvalidPin(Pin(3));

Or we can wrap errors generated by a function that returns std::io::Result using Rust's usual error propagation operator. For example from within a function that returns a GPIOResult (defined above) we can do the following:

let f = File::open("/dev/gpiomem")?;